pyproject.toml
[tool.poetry]
name = "cloud-secrets"
version = "0.1.0"
description = "A cloud-agnostic secrets management library"
authors = ["support <support@redacto.io>"]
readme = "README.md"
packages = [{include = "cloud_secrets"}]

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.36.20"
google-cloud-secret-manager = "^2.23.0"
azure-keyvault-secrets = "^4.9.0"
setuptools = "^75.8.0"
azure-identity = "^1.20.0"
django-environ = "^0.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
black = "^25.1.0"
isort = "^6.0.0"
flake8 = "^7.1.1"
mypy = "^1.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=cloud_secrets"
testpaths = ["tests"]
pythonpath = ["."]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88